Added permission check for clear, improved regexp and mode for MANCHE

+++ start.js
@@ -3,6 +3,7 @@ const discord = require('discord.js');
 const log = require('winston');
 require('dotenv').config({ path: 'keys.env' });
 const token = process.env.BOT_TOKEN;
+let mancheMode = false;
 
 // Configure logger settings
 log.remove(log.transports.Console);
@@ -21,15 +22,25 @@ bot.on('ready', () => {
 bot.on('message', (message) => {
 	log.info(`${message.author.username} - ${message.content}`);
 	let result;
-	const mancheRegEx = new RegExp('(?:.*)(?:di(?:s|t)?)(.*)', 'gi');
-	if (message.content === "$clear") {
+	const canClear = message.member.permissions.has("MANAGE_MESSAGES");
+	//add word bound with \b before di... group
+	const mancheRegEx = new RegExp('(?:.*)\\b(?:di(?:(?:s|t)[\\s-])?)(.*)', 'gi');
+	if (message.content === "$clear" && canClear) {
 		message.channel.fetchMessages({ limit: 100 }).then(messages => {
 			deleteAllNotPinned(messages);
 		}).catch(e => log.error(e));
-	} else if ((result = mancheRegEx.exec(message.content)) != null) {
+	} else if ((result = mancheRegEx.exec(message.content)) != null && mancheMode) {
 		log.info(`Matched message, gonna write ${result[1]}`);
 		message.channel.send(result[1].trim());
+	} else if (message.content === "$mancheMode") {
+		if (mancheMode) {
+			message.channel.send("Deactivating MANCHE mode");
+		} else {
+			message.channel.send("Activating MANCHE mode");
+		}
+		mancheMode = !mancheMode;
 	}
+	log.info(canClear);
 });
 
 bot.login(token);
